✏️ WOW-39 rename : 폴더명 및 폴더구조 수정
