+++ src/app/[locale]/(main)/my-page/page.tsx
